Reverse order of int and float

+++ src/covjson_pydantic/domain.py
@@ -56,9 +56,9 @@ class DomainType(str, Enum):
 
 
 class Axes(CovJsonBaseModel):
-    x: Optional[Union[ValuesAxis[float], ValuesAxis[int], ValuesAxis[str], CompactAxis]] = None
-    y: Optional[Union[ValuesAxis[float], ValuesAxis[int], ValuesAxis[str], CompactAxis]] = None
-    z: Optional[Union[ValuesAxis[float], ValuesAxis[int], ValuesAxis[str], CompactAxis]] = None
+    x: Optional[Union[ValuesAxis[int], ValuesAxis[float], ValuesAxis[str], CompactAxis]] = None
+    y: Optional[Union[ValuesAxis[int], ValuesAxis[float], ValuesAxis[str], CompactAxis]] = None
+    z: Optional[Union[ValuesAxis[int], ValuesAxis[float], ValuesAxis[str], CompactAxis]] = None
     t: Optional[ValuesAxis[AwareDatetime]] = None
     composite: Optional[ValuesAxis[Tuple]] = None
